% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append2csv.R
\name{append2csv}
\alias{append2csv}
\title{Append to CSV}
\usage{
append2csv(data_frame, file = "", append = TRUE, sep = ",",
  col.names = NULL, row.names = FALSE, ...)
}
\arguments{
\item{data_frame}{The object to be written, preferably a data frame or a matrix.}

\item{file}{Either a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{...}{Additional arguments. Same as for write.table.}
}
\description{
Wrapper to function write.table.
}
\details{
If the file doesn't exist, the csv will be created using col.names as header. If the file already exists, the data.frame will be appended without header.
}
\seealso{
\code{\link{write.table}}
}
