% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTempDir.R
\name{createTempDir}
\alias{createTempDir}
\title{Create Temporary Directory}
\usage{
createTempDir(parent_dir, temp_dir_prefix = "", length_random_suffix = 10)
}
\arguments{
\item{parent_dir}{Location where the temporary direcotiry will be created.}

\item{temp_dir_prefix}{Optional. String that will be used as prefix for the temporary directory.}

\item{length_random_suffix}{Optional. Number of random chars that will be used as suffix. By default 10.}
}
\description{
Create a temporary directory with a random name in a specific location.
}
\examples{
temporary_folder = createTempDir(temp_folder,"raster2psql")
on.exit(unlink(temporary_folder,recursive = TRUE),add=TRUE)
}

