% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_log.R
\name{debug_log}
\alias{debug_log}
\title{Prints Debug Log}
\usage{
debug_log(..., s = " ", c = " ")
}
\arguments{
\item{...}{Strings, or vectors. Same behavior as paste(...).}

\item{s}{Optional. A string to separate the terms. Same behavior as 'sep' on paste().}

\item{c}{Optional. A string to separate the collapsed vectors. Same behavior as 'collapse' on paste().}
}
\description{
Prints on the screen a string (or a vector) for debugging, indicating date and time of the system.
}
\details{
debug_log is enable by default. It could be disable using: options('reutiles-DEBUG_LOG'=FALSE)
}

