% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorIfNotType.R
\name{errorIfNotType}
\alias{errorIfNotType}
\title{Throw Error if Object is Wrong Type}
\usage{
errorIfNotType(x, test_type)
}
\arguments{
\item{x}{Logical expression.}

\item{test_type}{String containing the type to test. Valid types: 'character', 'numeric' or 'logical'}
}
\description{
Throw an error if object doesn't match the specified type
}
