% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{is.DateTime}
\alias{is.DateTime}
\title{Test of DateTime}
\usage{
is.DateTime(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Test if an object is of POSIXct class
}
\seealso{
\code{\link{POSIXct}}
}

