% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMainRscript.R
\name{isMainRscript}
\alias{isMainRscript}
\title{Check main Rscript}
\usage{
isMainRscript()
}
\description{
Return a boolean depending on whether the current script is the main script called from Rscript or not.
}

