% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isRscript.R
\name{isRscript}
\alias{isRscript}
\title{Check Rscript}
\usage{
isRscript(returnPath = FALSE)
}
\arguments{
\item{returnPath}{If FALSE (default), it returns TRUE if the script was called from Rscript, or FALSE if it was not.

When returnPath is TRUE, instead of returning a boolean, it returns the path of the script or NULL.}
}
\description{
Return a boolean depending on whether the script was called from Rscript or not.
}

