% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidFunctions.R
\name{isValidDate}
\alias{isValidDate}
\title{Check Date String Format}
\usage{
isValidDate(date, format = "\%Y-\%m-\%d")
}
\arguments{
\item{format}{By default equals to "\%Y-\%m-\%d"}
}
\description{
Return a boolean depending on whether the date string is in the right format or not.
}

