% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidFunctions.R
\name{isValidDateTime}
\alias{isValidDateTime}
\title{Check Datetime String Format}
\usage{
isValidDateTime(datetime, format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{format}{By default equals to "\%Y-\%m-\%d \%H:\%M:\%S"}
}
\description{
Return a boolean depending on whether the datetime string is in the right format or not.
}
