% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidFunctions.R
\name{isValidTime}
\alias{isValidTime}
\title{Check Time String Format}
\usage{
isValidTime(date, format = "\%H:\%M:\%S")
}
\arguments{
\item{format}{By default equals to "\%H:\%M:\%S"}
}
\description{
Return a boolean depending on whether the time string is in the right format or not.
}

