% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTempDir.R
\name{runInTempDir}
\alias{runInTempDir}
\title{Run Expression In a Temporary Directory}
\usage{
runInTempDir(expr, tmp_dir = tempdir(), prefix = "temp")
}
\arguments{
\item{expr}{Expression to be evaluated in a temporary directory.}

\item{tmp_dir}{Optional. Path to temporary main directory. Default tempdir()}

\item{prefix}{Optional. String that will be used as prefix for the temporary subfolder.}
}
\value{
Returns invisibly the last value evaluated in the expression.
}
\description{
Run Expression In a Temporary Directory
}
\note{
Working directory will be change to temporary subfolder. Once the evaluation of
      the expression is finished, the working directory will be restored and the
      temporary subfolder (with its contents) will be deleted. The temporary main
      directory will remain untouched.
}
\examples{
temp_wd = runInTempDir({getwd()})
temp_wd
}
