% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopStatus.R
\name{stopStatus}
\alias{stopStatus}
\title{Stop With Custom Status}
\usage{
stopStatus(..., status = 1)
}
\arguments{
\item{...}{zero or more objects. Same behavior as stop(...).}

\item{status}{Optional. The (numerical) error status to be returned to the operating system.}
}
\description{
stopStatus stops execution of the current expression and executes an error action returning a custom error status.
}
\details{
The default error handler for non-interactive use returns error status 1. Error status 2 is used for R 'suicide', that is a catastrophic failure, and other small numbers are used by specific ports for initialization failures. It is recommended that users choose statuses of 10 or more.
}

