% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressThisWarning.R
\name{suppressThisWarning}
\alias{suppressThisWarning}
\title{Suppress This Warning}
\usage{
suppressThisWarning(expr, pattern)
}
\arguments{
\item{expr}{Expression to be evaluated.}

\item{pattern}{Character string (regexp) to be matched with the possible warnings.}
}
\description{
Suppress a specific warning
}
