% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisScript.R
\name{thisScript}
\alias{thisScript}
\title{Get Path of the Current Script}
\usage{
thisScript()
}
\value{
String path or NULL
}
\description{
Returns the path of the current script. When the path couldn't be found, it returns NULL (e.g. R console).
}
