% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{as.DateTime}
\alias{as.DateTime}
\title{Date-time Conversion}
\usage{
as.DateTime(x, tz = "UTC", origin = "1970-01-01", ...)
}
\arguments{
\item{x}{An object to be converted.}

\item{tz}{A time zone specification to be used for the conversion. Default 'UTC'.}

\item{origin}{a date-time object, or something which can be coerced to. Default '1970-01-01'.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\description{
Wrapper to function as.POSIXct
}
\seealso{
\code{\link{as.POSIXct}}
}
