% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTempDir.R
\name{createTempDir}
\alias{createTempDir}
\title{Create Temporary Directory}
\usage{
createTempDir(parent_dir, temp_dir_prefix = "", length_random_suffix = 10,
  trailing_slash = TRUE)
}
\arguments{
\item{parent_dir}{Location where the temporary direcotiry will be created.}

\item{temp_dir_prefix}{Optional. String that will be used as prefix for the temporary directory.}

\item{length_random_suffix}{Optional. Number of random chars that will be used as suffix. By default 10.}

\item{trailing_slash}{Optional. Boolean indicating if a trailing slash should be added. By default TRUE.}
}
\description{
Create a temporary directory with a random name in a specific location.
}
\examples{

some.function <- function(..., temp_folder = '/default/path'){
  temporary_folder = createTempDir(temp_folder,"prefix_str")
  print(temporary_folder)
  # this will remove the temporary folder at the end of function execution
  on.exit(unlink(temporary_folder,recursive = TRUE),add=TRUE)

  # Some code

}
}
